+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 ### Changed
+- `nexus` repository description
 - show missing variables warning message only if `publish` task is called
 
 ### Deprecated